Initial support for reading .dsk images.
For each sector we compute a sha1 hash, and use this to recognie some common sector types (zero, common RWTS sectors)

Also compute the zlib compressibility of the sector as a way of estimating entropy, which might be useful in providing insight into whether it contains e.g. code or data.