Add a container.Container() type that carries a list of Anomaly() objects associated to the container, and builds a tree of child containers with support for depth-first recursion
Convert all of the disk container classes to use Container() and
register themselves with this tree.  This constructs a hierarchy of
container objects rooted in Disk that describe the structures found on
the disk image.

Convert most of the print statements and some of the assertions to
register Anomaly records instead

Fix/improve some of the __str__ representations